+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Change Log
 
-## 04 Jul 25
+## 2025-07-09
+
+### Added
+
+- Wrap the Google Socialite callback route in the `web` middleware
+- Update README.md with information about the `web` middleware requirement
+
+### Changed
+
+- Remember Google Socialite login
+
+## 2025-07-04
 
 - Add push notification biolerplate
 
-## 03 Jul 25
+## 2025-07-03
 
 ### Added